MVP!
*removed all old folder from repo
* broke up styles into seperate files in make offer
* finished back card (don't look great but, owner signed off on it)
