--- old/Menu.css
@@ -1,50 +1,0 @@
-#flyoutMenu {
-  width: 100vw;
-  height: 100vh;
-  background-color: #1565c0;
-  position: fixed;
-  top: 0;
-  left: 0;
-  transition: transform .3s
-  cubic-bezier(0, .52, 0, 1);
-  overflow: scroll;
-  z-index: 1000;
-}
- 
-.hide {
-  transform: translate3d(-100vw, 0, 0);
-}
- 
-.show {
-  transform: translate3d(0vw, 0, 0);
-}
-
-
-.flyoutLinks {
-  padding: 28px 28px 5px 28px;
-  text-align: center;
-  font-size: 2rem;
-  font-family: 'Roboto', sans-serif;
-  letter-spacing: .22rem;
-  color: white;
-  text-decoration: none;
-  transition: all .20s ease-in;
-  margin: auto;
-  font-variant: small-caps;
-  border-color: #1565c0
-}
-.flyoutLinks:hover {
-  transform: scale3d(1.2,1.3,1);
-}
-.flyoutActive {
-  border-bottom: 8px solid #eb1c24;
-}
-/* @media(min-width: 550px) {
-  .flyoutLinks {
-    font-size: 2rem;
-  }
-}  */
-/* .flyoutLinks:hover {
-  border-bottom: 8px solid #eb1c24;
-} */
-
