--- old/MenuButtonStyles.css
@@ -1,163 +1,0 @@
-
-/* #roundButton {
-  float: right;
-    background-color: #96D9FF;
-    width: 50px;
-    height: 50px;
-    border-radius: 50%;
-    border: 10px solid #0065A6;
-    outline: none;
-    transition: all .2s cubic-bezier(0, 1.26, .8, 1.28);
-    margin-top: 5px;
-    margin-left: -130px;
-    margin-right: 15px;
-} */
-
-/* @media (min-width: 768px) {
-  #roundButton {
-    margin-top: 20px;
-    margin-bottom: 20px;
-    margin-left: -130px;
-    margin-right: 50px;
-  }
-}
-
-@media (min-width: 1022px) {
-  #roundButton {
-    display: none;
-  }
-} */
-
-
-/* #roundButton:hover {
-  background-color: #96D9FF;
-  cursor: pointer;
-  border-color: #003557;
-  transform: scale(1.2, 1.2);
-}
- 
-#roundButton:active {
-  border-color: #003557;
-  background-color: #FFF;
-} */
-
-/* .menu {
-  float: right;
-  cursor: pointer;
-  outline: none;
-  background: transparent;
-  border: none;
-  outline: none;
-  font-size: 0;
-  overflow: visible;
-  margin: 30px;
-  transition-property: transform;
-  transition-duration: .5s;
-  
-
-  margin-top: 5px;
-  margin-left: -130px;
-  margin-right: 15px;
-} */
-
-/* #roundButton {
-    width: 50px;
-    height: 50px;
-    transition: all .2s cubic-bezier(0, 1.26, .8, 1.28);
-} */
-
-/* @media (min-width: 768px) {
-.menu {
-  margin-top: 20px;
-  margin-bottom: 20px;
-  margin-left: -130px;
-  margin-right: 50px;
-}
-}
-
-@media (min-width: 1022px) {
-  .menu {
-  display: none;
-}
-
-} */
-
-/* #roundButton {
-  float: right;
-    background-color: #96D9FF;
-    width: 50px;
-    height: 50px;
-    transition: all .2s cubic-bezier(0, 1.26, .8, 1.28);
-    margin-top: 5px;
-    margin-left: -130px;
-    margin-right: 15px; */
-
-
-
-.menu-lines {
-  
-  position: relative;
-  display: inline-block;
-  width: 47px;
-  height: 37px;
-  vertical-align: middle;
-}
-
-.menu-lines span {
-  position: absolute;
-  display: block;
-  height: 6px;
-  width: 70%;
-
-}
-@media(min-width: 768px) {
-  .menu-lines span {
-    height: 8px !important;
-    width: 100% !important;
-  }
-  .menu-lines span:nth-child(2) {
-    top: 15px !important;
-  }
-  .menu-lines span:nth-child(3) {
-    top: 30px !important;
-  }
-}
-.menu-lines span:nth-child(1) {
-  top: 0
-}
-
-.menu-lines span:nth-child(2) {
-  top: 10px
-}
-
-.menu-lines span:nth-child(3) {
-  top: 20px
-}
-.bar1 {
-  background: #eb1c24;
-}
-.bar2 {
-  background-color: #1565c0;
-}
-.bar3 {
-  background: #eb1c24;
-}
-.bar1, .bar2, .bar3 {
-  width: 35px;
-  height: 5px;
-  
-  margin: 6px 0;
-  transition: 0.4s;
-}
-
-.change .bar1 {
-  -webkit-transform: rotate(-45deg) translate(-9px, 6px) ;
-  transform: rotate(-45deg) translate(-12px, 12px) ;
-}
-
-.change .bar2 {opacity: 0;}
-
-.change .bar3 {
-  -webkit-transform: rotate(45deg) translate(-8px, -8px) ;
-  transform: rotate(45deg) translate(-8px, -8px) ;
-}
