--- old/app.css
@@ -1,18 +1,0 @@
-@keyframes fadeIn {
-  from {
-    opacity: 0;
-  }
-
-  to {
-    opacity: 1;
-  }
-}
-
-.fadeIn {
-  animation-name: fadeIn;
-}
-
-.animated {
-  animation-duration: .5s;
-  animation-fill-mode: both;
-}
