--- old/blog.txt
@@ -1,27 +1,0 @@
-<h1>Blog Component</h1>
-      <p>
-      Imagine climbing into your bed after a hard day and being so
-      comfortable that you drift to sleep, and stay asleep through
-      the night. Now picture waking in the morning feeling rested,
-      rejuvenated and ready to take on the day. THAT, is what the 
-      right mattress for you can do!
-      </p>
-      <p>Did you know that your mattress is the most important thing you purchase for your health?</p>
-
-      <p>Think about it.</p>
-      <p>You spend a third of your life in bed, and that third
-        effects all aspects of the other two-thirds.
-        If you wake up in pain, groggy, exhausted, etc., your entire 
-        day is ruined, your productivity decreases and the chance of
-        getting hurt or sick goes up. Imagine getting your best 
-        night’s sleep every night, waking up feeling refreshed, 
-        well rested and ready for the day.
-        If you toss and turn, you need a new mattress. It’s really 
-        that simple.
-</p>
-<p>Tossing and turning means that the mattress you are on is not giving you the comfort, pressure relief, or support that your body needs. We move around to try and get comfortable and every time we move, we wake up to some degree. We come out of REM sleep, which is the level of sleep that is most important to waking up feeling rested. The longer we stay in REM sleep at a time, the better we feel in the morning.</p>
-<p>We toss and turn for numerous reasons, the most common is the firmness level of your mattress. It could be that your mattress’s comfort layer is too firm and putting pressure on your shoulders and hips, as well as cutting off circulation to your internal organs. These reasons cause our bodies to make us move every hour or so. Sometimes more often. That means you are only getting an hour or less of REM sleep at a time. 
-If you have the correct comfort in your mattress, you will stay in REM sleep longer and wake up feeling more rested, refreshed and ready for your day.</p>
-
-<p>Remember, the firmness or softness of your mattress has no effect on the support for your spine. The way mattresses are made today, from the bottom up, means that you get your support layer underneath and then the comfort layer on top. It’s not like the old days, when you had to suffer an uncomfortably firm mattress to get spinal support. Now you can get whatever softness is the most comfortable to you and still get the same spinal alignment. Win. Win.</p>
-<p>There is NO one size fits all mattress. It doesn’t exist. Everyone’s body is different and has it’s own needs. Never buy a mattress you can’t lay on first. Come in and let our Non-Commission Sales Expert comfort test you and fit you for the mattress that is right for your body.</p>
