--- old/home.css
@@ -1,56 +1,0 @@
-.slide {
-  margin-bottom: 0px !important;
-  margin-top:  0px !important;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-around;
-}
-
-.carousel .slide {
-  background: white;
-  margin: auto;}
-/* }
-.slider {
-  margin: auto !important;
-}
-.carousel {
-  margin: auto;
-}
-.slider-wrapper {
-  margin: auto;
-  width: auto;
-}
-.control-arrow .control-prev {
-  margin: auto;
-
-} */
-.control-prev {
-
-  width: 50px !important;
-  font-size: 2rem;
-}
-.control-next {
-  width: 50px;
-}
-
-@media(min-width:1100px) {
-  .control-next {
-    width: 70px;
-  }
-  .control-prev {
-    width: 70px !important;
-    font-size: 2rem;
-  }
-}
-
-.carousel .control-arrow:before, .carousel.carousel-slider .control-arrow:before {
-    border-top: 20px solid transparent;
-    border-bottom: 20px solid transparent;
-}
-.carousel .control-next.control-arrow:before {
-  border-left: 20px solid #fff;
-}
-
-.carousel .control-prev.control-arrow:before {
-  border-right: 20px solid #fff;
-}
