--- old/sharing.js
@@ -1,33 +1,0 @@
-<div>
-<title>E.S.C. Mattress Center</title>
-<meta name="description" content="A licensed mattress retailer for Sealy, Stearns and Foster, and Tempur-Pedic offering almost fifty mattresses with prices to fit every budget.  ESC Mattress Center is a locally owned company in Everett WA with more than twenty years of mattress industry experience between our non-commissioned staff."/>
-<meta property="og:type"   content="website" />
-<meta property="og:site_name" content="E.S.C. Mattress Center"/>
-<meta property="og:url" content="https://www.escmattresscenter.com/"/>
-<meta property="og:image:secure_url" content="https://www.escmattresscenter.com/static/media/logo.952ae51d.png"/> 
-<meta property="og:image:width" content="1200"/>
-<meta property="og:image:height" content="627"/>
-<meta property="og:image:alt" content="E.S.C Mattress Center's logo of a panda"/>
-<meta property="og:title" content="E.S.C. Mattress Center"/>
-<meta property="og:description" content="Sleep like the experts do."/>
-
-
-{/* <meta property="og:url" content={`https://www.escmattresscenter.com/adjustable/${base.uri}`}/> */}
-{/* <meta property="og:url" content={`https://www.escmattresscenter.com/blog/${post.slug}`}/> */}
-{/* <meta property="og:url" content="https://www.escmattresscenter.com/blog"/> */}
-{/* <meta property="og:url" content="https://www.escmattresscenter.com/brands"/> */}
-{/* <meta property="og:url" content="https://www.escmattresscenter.com/financing"/> */}
-{/* <meta property="og:url" content="https://www.escmattresscenter.com/brands/sealy"/> */}
-{/* <meta property="og:url" content="https://www.escmattresscenter.com/brands/stearns"/> */}
-  {/* <meta property="og:url" content="https://www.escmattresscenter.com/brands/tempurpedic"/> */}
- {/* <meta property="og:url" content={`https://www.escmattresscenter.com/brands/${name}/${mattress.uri}`}/> */}
-  // let name = '';
-    // if (mattress.brandName === 'Tempur-PEDIC') {
-    //   name = 'tempurpedic';
-    // } else if (mattress.brandName === 'Sealy'){
-    //   name = 'sealy'
-    // } else {
-    //   name = 'stearns'
-    // }
-
-</div>
