--- old/sitemap-builder.js
@@ -1,10 +1,0 @@
-require('babel-register');
-
-const router = require('./router').default;
-const Sitemap = require('../').default;
-
-(
-	new Sitemap(router)
-		.build('http://www.escmattresscenter.com/')
-		.save('./sitemap.xml')
-);
