--- old/sitemap.xml
@@ -1,251 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
-<url>
-  <loc>https://www.escmattresscenter.com</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/financing/</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/about</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/blog</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/adjustable</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/adjustable/TEMPUR-Ergo-Premier</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/adjustable/Stearns&amp;Foster-Reflexion7</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/adjustable/TEMPUR-Ergo-Plus</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/adjustable/Sealy-Ease</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/accessories</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/brands</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/tempurpedic</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/tempurpedic/TEMPUR-Contour-Supreme</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/tempurpedic/TEMPUR-Contour-Elite</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/tempurpedic/TEMPUR-Contour-Elite-Breeze</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/tempurpedic/TEMPUR-Contour-Rhapsody-Luxe</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/tempurpedic/TEMPUR-Contour-Rhapsody-Luxe-Breeze</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/tempurpedic/TEMPUR-Cloud-Prima</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/tempurpedic/TEMPUR-Cloud-Supreme</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/tempurpedic/TEMPUR-Cloud-Supreme-Breeze</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/tempurpedic/TEMPUR-Cloud-Elite</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/tempurpedic/TEMPUR-Cloud-Luxe</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/tempurpedic/TEMPUR-Cloud-Luxe-Breeze</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/tempurpedic/TEMPUR-Flex-Prima</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/tempurpedic/TEMPUR-Flex-Supreme</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/tempurpedic/TEMPUR-Flex-Supreme-Breeze</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/tempurpedic/TEMPUR-FLEX-Elite</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/tempurpedic/TEMPUR-Legacy-Limited-Edition</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/stearns/Estate-Oak-Terrace-Luxury-Firm</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/stearns/Estate-Oak-Terrace-Luxury-Plush</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/stearns/Estate-Oak-Terrace-Luxury-Firm-PT</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/stearns/Estate-Scarborough-UF</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/stearns/Estate-Scarborough-Luxury-Plush-PT</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/stearns/Lux-Estate-Rookwood-Luxury-Firm</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/stearns/Lux-Estate-Hybrid-Brooklet-Luxury-CF</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/stearns/Reserve-No.3-Luxury-Firm-PT</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/sealy/</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/sealy/Essentials-Joyous-Firm</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/sealy/Essentials-Merry-Firm</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/sealy/Essentials-Merry-Plush</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/sealy/Essentials-Happiness-Cushion-Firm</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/sealy/Essentials-Happiness-Plush</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/sealy/Essentials-Happiness-Plush-PT</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/sealy/Conform-Essentials-Upbeat-Firm</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/sealy/Conform-Essentials-Treat-Cushion-Firm</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/sealy/Conform-Essentials-Optimistic-Plush</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/sealy/To-Go-12-Plush-MemFoam-Bedin-aBox</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/sealy/Conform-Performance-High-Spirits-Firm</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/sealy/Conform-Performance-Fondness-CF</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/sealy/Conform-Performance-Thrilled-Plush</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/sealy/Response-Performance-Merriment-CF</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/sealy/Response-Performance-Merriment-Plush</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/sealy/Response-Performance-Merriment-Cushion-Firm-PT</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/sealy/Response-Performance-Surprise-Firm</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/sealy/Response-Performance-Surprise-Plush-PT</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/sealy/Response-Premium-Satisfied-CF</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/sealy/Response-Premium-Satisfied-Plush</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/sealy/Response-Premium-Exuberant-UF</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/sealy/Response-Premium-Exuberant-Plush-PT</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/sealy/Hybrid-Cobalt-Firm</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/sealy/Hybrid-Copper-Cushion-Firm</loc>
-  <priority>0.5</priority>
-</url>
-<url>
-  <loc>https://www.escmattresscenter.com/sealy/Hybrid-Copper-Plush</loc>
-  <priority>0.5</priority>
-</url>
-</urlset>
