+++ src/Components/Price/MakeOffer/Styles/ModalStyles.jsx
@@ -1,5 +1,16 @@
 import styled from 'styled-components';
 
+export const ModalContainer = styled.div`
+  position: fixed;
+  height: 100%;
+  width: 100%;
+  top: 0;
+  left: 0;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+`;
+
 export const CardContainer = styled.div`
   position: relative;
   width: 90vw; 
@@ -46,70 +57,22 @@ export const CardContainer = styled.div`
   }
   @media(min-width: 931px) and (min-height: 920px) {
     width: 50vw;
-    height: 50vh;
+    height: 57vh;
   }
   @media(min-width: 1024px) and (min-height: 920px) and (orientation: portrait) {
-    width: 40vw;
-    height: 44vh;
+    width: 46vw;
+    height: 41vh;
   }
   @media(min-width: 1024px) and (min-height: 920px) and (orientation: landscape){
     width: 500px;
     height: 530px;
   }
- 
-  /* @media(min-width: 1500px) and (min-height: 920px) {
-    width: 35vw;
-    height: 60vh;
-  } */
-  /* @media(min-width: 375px) and (min-height: 920px) and (orientation: portrait){
-    height: 53vh;
-  }
-  @media(min-width: 420px) and (min-height: 920px) and (orientation: portrait){
-    height: 51vh;
-  }
-  @media(min-width: 540px) and (min-height: 920px) and (orientation: portrait){
-    width: 60vw;
-  }
-  @media(min-width: 768px) and (min-height: 920px) and (orientation: portrait){
-    height: 60vh;
-  }
-  @media(min-width: 857px) and (min-height: 920px) and (orientation: portrait){
-    height: 57vh;
-  } */
- 
-  
 `;
-export const TextArea = styled.textarea`
-  outline: none;
-  padding: 7px;
-  background: transparent;
-  border: none;
-  margin-bottom: 10px;
-  border-bottom: 1px solid #ddd;
-  transition: all .2s ease-in-out;
-  resize:none;
-  overflow: hidden;
-  height: 60px;
-  @media(orientation: landscape) {
-    height: auto;
-    margin-bottom: 5px;
-  }
-  @media(orientation: landscape) and (min-width:640px){
-    margin-bottom: 1px;
-    }
-  @media (min-width: 730px) and (orientation: landscape) {
-    margin-bottom: 10px;
-  }
-  @media (min-width: 812px) and (orientation: landscape) {
-    margin-bottom: 5px;
-  }
-  @media (min-width: 768px) and (min-height:823px) and (orientation: portrait) {
-    font-size: 1.2rem;
-  }
-  @media(min-width: 1024px) {
-    font-size: 1.2rem;
-    height: 70px;
-    margin-bottom: 10px;
-  }
 
+export const Card = styled.div`
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  transform-style: preserve-3d;
+  transition: all 0.5s ease;
 `;
